pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ehs-ai"
version = "0.3.1"
description = "Environmental Health & Safety incident assistant powered by pydantic-ai."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "EHS AI Team" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.111",
    "openai>=1.40",
    "pydantic>=2.7",
    "pydantic-ai>=0.0.8",
    "pydantic-settings>=2.3",
    "langgraph>=1.0.3",
    "langgraph-cli>=0.4.7",
    "chromadb>=1.3",
    "sentence-transformers>=3.0",
    "pillow>=10.4",
    "fpdf2>=2.7",
    "python-dotenv>=1.0",
    "uvicorn>=0.30"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5",
    "mypy>=1.10",
    "pytest>=8.2"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ehs_ai*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

